Fix DTEND wit type DATE
If DTEND uses the type DATE instead of DATETIME the rule that end has to be after start still applies. Therefore DTEND always must be at least one day after DTSTART, even if the appointment only last one day. As a consequence DTEND has to be the first day _after_ the appointment.

Addiotionally add handling for appointments starting and ending in different months.